requirements.txt
python-dotenv==1.0.1
chromadb==0.5.11
langchain==0.3.7
langchain-community==0.3.7
openai==1.54.3
ollama==0.3.3
pandas==2.2.3
PyPDF2==3.0.1
python-docx==1.1.2
sentence-transformers==3.2.1
tqdm==4.66.5
